Fixed prerendering page error
